initial version of reading a writing ipynb files
To give it a spin right now, just CD to the directory of these files,
and run "make". This will open up a new buffer for every cell in (and
output) in the notebook.

To test what currently gets written out from these buffers in the even
that you want to combine them back and make a full notebook, run 'make
diff'. This will do the above, save a new .ipynb from those buffers, and
then open up vimdiff allowing you to compare new.ipynb to the original
tmp.ipynb

WARNING: this process is lossy at the moment, use at your own risk!